Change programm 076 so that once the user has completed their list of
names, display the full list and ask them to type in one of the
names on the list. Display the position of that name in the list.
Ask the user if they still want that person to come to the party.
If they answer "no", delete that entry from the list and display
the list again.