Import the data from the Books.csv file into a list. Display the list
to the user. Ask them to select which row from the list they want
to delete and remove it from the list. Ask the user which data they
want to change and allow them to change it. Write the data back to
the original.csv file, overwriting the existing data with the
amended data.