Display the following menu to the user:
1) Addition
2) Subtraction
Enter 1 or 2:
If they enter a 1, it should run a subprogram that will generate two
random numbers between 5 and 20, and ask the user to add them together.
Work out the correct answer and return both the user's answer and the
correct answer.
If they entered 2 as their selection on the menu, it should run a
subprogram that will generate one number between 25 and 50 and another
number between 1 and 25 and ask them to work out num1 minus num2. This
way they will not have out num1 minus num2. This way they will not have
to worry about negative answers.
Create another subprogram that will check if the user's answer mathches
the actual answer. If it does, display "Correct", otherwise display a
message that will say "Incorrect, the answer is" and display the real
answer.
If they do not select a relevant option on the first menu you should
display a suitable message.