In Python, in is not technically possible to directly delete a
record from a .csv file. Instead you need to save the file to a
temporary list in Python, make the changes to the list and then
overwrite the original file with the temporary list.
Change the previous program to allow you to do this Your menu
should now look like this:
1) Add to file
2) View all records
3) Delete a record
4) Quit program
Enter the number of your selection: