Create a program that will ask the user to enter a number in a box.
When they click on a button it will add that number to a total and
display it in another box. This can be repeated as many times as they
want and keep adding to the total. There should be another button that
resets the total back to 0 and empties the original text box, ready for
them to start again.