Create a new program that will generate two random whole numbers
between 10 and 50. It should ask the user to add the numbers together
and type in the answer. If they get the question correct, display a
suitable image such as a tick; if they get the answer wrong, display
another suitable image such as a cross. They should click on a Next
button to get another question.