Create a table called employees which has four fields(id, name, dept
and salary). It specifies the data type for each field, defines which
field is the primary key and which fields cannot be left blank. The
triple speech marks allow the code to be split over several lines to
make it easier to read rather than having it all displayed in one line.