Sqlite3. Using the PhoneBook database from program 139, write a
program that will display the following menu.
Main Menu:
1) View phone book
2) Add to phone book
3) Search for surname
4) Delete person from phone book
5) Quit
Enter your selection:
If the user selects 1, they should be able to view the entire
phonebook. If they select 2, it should allow them to add a new
person to the phonebook. If they select 3, it should ask them for a
surname and then display only the records of people with the same
surname. If they select 4, it should ask for an ID and then delete
that record from the table. If they select 5, it should end the
program. Finally, it should display a suitable message if they enter
an incorrect selection from the menu.
They should return to the menu after each action, until they select 5.