Sqlite3. Using the Bookinfo database from program 141, display the
list of authors and their place of birth. Ask the  user to enter a
place of birth. Ask the user to enter a place of birth and then show
the title, date published and author's name for all the books by
authors who were born in the location they selected.